requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
djangorestframework==3.7.7
jupyter==1.0.0
PyJWT==1.6.1
gunicorn==19.8.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4